Move FAO to content as doesnt work in embed
