+++ tests/Discord/FlowMeasure/Message/FlowMeasureWithdrawnMessageTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Discord\FlowMeasure\Message;
 
-use App\Discord\FlowMeasure\Description\EventNameAndInterestedParties;
+use App\Discord\FlowMeasure\Description\EventName;
 use App\Discord\FlowMeasure\Message\FlowMeasureWithdrawnMessage;
 use App\Discord\Message\Embed\Colour;
 use App\Models\DiscordTag;
@@ -12,12 +12,17 @@ use Tests\TestCase;
 
 class FlowMeasureWithdrawnMessageTest extends TestCase
 {
-    public function testItHasNoContent()
+    public function testItHasFaoContent()
     {
+        $fir = FlightInformationRegion::factory()->has(DiscordTag::factory()->count(2))->create();
         $measure = FlowMeasure::factory()->create();
+        $measure->notifiedFlightInformationRegions()->sync([$fir->id]);
 
         $this->assertEquals(
-            '',
+            sprintf(
+                "**FAO**: %s\nPlease acknowledge receipt with a :white_check_mark: reaction.",
+                $fir->discordTags->pluck('tag')->map(fn(string $tag) => '<' . $tag . '>')->join(' ')
+            ),
             (new FlowMeasureWithdrawnMessage($measure))->content()
         );
     }
@@ -36,7 +41,7 @@ class FlowMeasureWithdrawnMessageTest extends TestCase
                 [
                     'title' => $measure->identifier . ' - ' . 'Withdrawn',
                     'color' => Colour::WITHDRAWN->value,
-                    'description' => (new EventNameAndInterestedParties($measure))->description(),
+                    'description' => (new EventName($measure))->description(),
                     'fields' => [
                         [
                             'name' => 'Minimum Departure Interval [MDI]',
@@ -75,7 +80,7 @@ class FlowMeasureWithdrawnMessageTest extends TestCase
                 [
                     'title' => $measure->identifier . ' - ' . 'Withdrawn',
                     'color' => Colour::WITHDRAWN->value,
-                    'description' => (new EventNameAndInterestedParties($measure))->description(),
+                    'description' => (new EventName($measure))->description(),
                     'fields' => [
                         [
                             'name' => 'Minimum Departure Interval [MDI]',
